Subtree update automation: use fast version of subtree split
repo-filter does not produce consistent SHA hashes. While we were aware
of this, we hadn't realized that this made incremental non-interactive
updates of the branch a matter of luck as git may or may not be able to
automatically resolve conflicts. (It worked for the update to
2025-02-10, but failed for the next one.)

git-subtree-split does produce consistent SHA hashes, but the native
`git-subtree-split` implementation is a shell script, and thus was found
to be too slow (which is why we chose repo-filter in the first place).

splitsh-lite is an implementation of just the `subtree split` command in
Go and libgit2. This makes `subtree split` even faster than
`repo-filter` at 35 seconds vs 50 seconds (on a GitHub runner).